% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abort.R, R/recover.R, R/suppress.R,
%   R/with_verbose.R
\name{abort_on_messages}
\alias{abort_on_messages}
\alias{abort_on_warnings}
\alias{recover_on_messages}
\alias{recover_on_warnings}
\alias{suppress_messages}
\alias{suppress_warnings}
\alias{with_verbose_messages}
\alias{with_verbose_warnings}
\title{Intercept messages and warnings}
\usage{
abort_on_messages(.expr, .f = NULL, ..., .classes = NULL, .packages = NULL)

abort_on_warnings(.expr, .f = NULL, ..., .classes = NULL, .packages = NULL)

recover_on_messages(.expr, .f = NULL, ..., .classes = NULL, .packages = NULL)

recover_on_warnings(.expr, .f = NULL, ..., .classes = NULL, .packages = NULL)

suppress_messages(.expr, .f = NULL, ..., .classes = NULL, .packages = NULL)

suppress_warnings(.expr, .f = NULL, ..., .classes = NULL, .packages = NULL)

with_verbose_messages(.expr, .f = NULL, ..., .classes = NULL, .packages = NULL)

with_verbose_warnings(.expr, .f = NULL, ..., .classes = NULL, .packages = NULL)
}
\arguments{
\item{.expr}{expression to evaluate}

\item{.f}{predicate function to apply on the message, a purrr lambda, or a
string containing a regular expression}

\item{...}{additional parameters to forward to \code{.f}}

\item{.classes}{condition classes to consider, by default everything qualifies}

\item{.packages}{packages to consider, by default everything qualifies}
}
\value{
The output of \code{.expr}
}
\description{
Intercept messages and warnings
}
\examples{
suppress_warnings(sqrt(-1), "foo")
suppress_warnings(sqrt(-1), "NaN")
}
